refactor: 파일명 index 변경
