+++ src/App.js
@@ -1,10 +1,10 @@
 import Header from "./components/Header.js";
-import TabBarView from "./components/TabBar/TabBar.js";
+import TabBarView from "./components/TabBar/index.js";
 import AddRestaurantModal from "./components/modal/AddRestaurantModal/index.js";
 import FilterBarManager from "./components/FilterBar/FilterBarManager.js";
 import TabManager from "./components/TabBar/TabManager.js";
 import RestaurantManager from "./components/RestaurantList/RestaurantManager.js";
-import TabBar from "./components/TabBar/TabBar.js";
+import TabBar from "./components/TabBar/index.js";
 
 class App {
   #$target;
